moved IconImage to different namespace
