+++ AnnoDesigner/PresetsLoader/IconLoader.cs
@@ -1,5 +1,5 @@
-﻿using AnnoDesigner.Presets;
-using AnnoDesigner.UI;
+﻿using AnnoDesigner.model;
+using AnnoDesigner.Presets;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
